requirements.txt
asgiref==3.8.1
dj-database-url==2.1.0
Django==5.0.6
djangorestframework==3.15.1
gunicorn==22.0.0
packaging==24.0
pillow==10.3.0
psycopg2==2.9.9
sqlparse==0.5.0
typing_extensions==4.11.0
tzdata==2024.1
whitenoise==6.6.0
